renamed misc package to terminal
