+++ src/terminal/terminalMsg.go
@@ -1,4 +1,4 @@
-package misc
+package terminal
 
 import (
 	"fmt"
